requirements.txt
requests
python-dotenv
pillow
streamlit
huggingface_hub
